--- lesson1/src/main/java/ru/otus/dao/IStudentTestingDao.java
@@ -1,10 +1,0 @@
-package ru.otus.dao;
-
-import ru.otus.domain.Question;
-
-import java.util.ArrayList;
-
-public interface IStudentTestingDao {
-
-    void startStudentTesting(ArrayList<Question> questionList);
-}
