+++ lesson1/src/main/java/ru/otus/service/StudentTestingService.java
@@ -2,5 +2,5 @@ package ru.otus.service;
 
 public interface StudentTestingService {
 
-    void startStudentTesting(ParserCsvService service, String filePath);
+    void startStudentTesting();
 }
