--- lesson1/src/main/resources/application.properties
@@ -1,2 +1,0 @@
-# csv-file path
-csv.path=lesson1/src/main/resources/questions.csv
