+++ lesson1/src/main/resources/spring-context.xml
@@ -3,19 +3,18 @@
        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
        xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd">
 
-    <bean id="studentTestDao" class="ru.otus.dao.StudentTestingDao">
-    </bean>
-
     <bean id="parseCsvDao" class="ru.otus.dao.ParserCsvDao">
     </bean>
 
-    <bean id="studentTestingService" class="ru.otus.service.StudentTestingServiceImpl">
-        <constructor-arg name="dao" ref="studentTestDao">
+    <bean id="parserCsvService" class="ru.otus.service.ParserCsvServiceImpl">
+        <constructor-arg name="dao" ref="parseCsvDao">
+        </constructor-arg>
+        <constructor-arg name="filePath" value="lesson1/src/main/resources/questions.csv">
         </constructor-arg>
     </bean>
 
-    <bean id="parseCsvService" class="ru.otus.service.ParserCsvServiceImpl">
-        <constructor-arg name="dao" ref="parseCsvDao">
+    <bean id="studentTestingService" class="ru.otus.service.StudentTestingServiceImpl">
+        <constructor-arg name="csvParser" ref="parserCsvService">
         </constructor-arg>
     </bean>
